include tests for both TransDecoder and TD2
